Move GRPC_STATUS_DETAILS_BIN_KEY to MetadataUtil
