+++ grpc/src/main/java/com/linecorp/armeria/common/grpc/GoogleGrpcStatusFunction.java
@@ -16,7 +16,7 @@
 
 package com.linecorp.armeria.common.grpc;
 
-import static com.linecorp.armeria.internal.common.grpc.GrpcMetadataKeys.GRPC_STATUS_DETAILS_BIN_KEY;
+import static com.linecorp.armeria.internal.common.grpc.MetadataUtil.GRPC_STATUS_DETAILS_BIN_KEY;
 import static java.util.Objects.requireNonNull;
 
 import com.linecorp.armeria.common.RequestContext;
